+++ example/ck_tile/18_hstu_attention/hstu_attention_jagged_forward_dispatch.hpp
@@ -126,9 +126,9 @@ struct jagged_forward_causal_local_bias_dropout_dispatch
         constexpr dim3 kBlockSize              = HstuKernel::BlockSize();
         constexpr ck_tile::index_t kBlockPerCu = HstuKernel::kBlockPerCu;
 
-        (void)ck_tile::launch_kernel(ck_tile::stream_config{stream, false},
-                                     ck_tile::make_kernel<kBlockSize.x, kBlockPerCu>(
-                                         HstuKernel{}, kGridSize, kBlockSize, 0, kargs));
+        (void)ck_tile::launch_kernel(
+            ck_tile::stream_config{stream, false},
+            ck_tile::make_kernel<kBlockPerCu>(HstuKernel{}, kGridSize, kBlockSize, 0, kargs));
     };
 };